Update desktop app structure and modify greeting message
The commit includes an update to the desktop app structure in which the references to the CheckMade.Desktop.App were updated. It also modifies the main greeting message in the MainViewModel. Additional modifications include removing certain project references in CheckMade.Mobile.Startup.csproj file and creating new files for the desktop app configuration.