Added CheckMade.Telegram.FunctionApp project to solution
A new project called CheckMade.Telegram.FunctionApp was created and added to the solution. This also includes the setup of initial project settings such as the target framework (.NET 8.0) and Azure Functions version (v4). In addition, a .gitignore file has been generated for the FunctionApp and host.json was added, configuring the function's runtime behavior.